requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
awsebcli==3.20.10
billiard==4.1.0
blessed==1.20.0
botocore==1.31.62
Brotli==1.1.0
celery==5.3.4
cement==2.8.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
cryptography==41.0.4
cssselect2==0.7.0
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
filelock==3.12.4
fonttools==4.43.0
html5lib==1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
lxml==4.9.3
mysqlclient==2.2.0
oscrypto==1.3.0
packaging==23.2
pathspec==0.10.1
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.16.2
pyphen==0.14.0
pypng==0.20220715.0
pytest==7.4.2
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
referencing==0.30.2
reportlab==3.6.13
requests==2.31.0
rpds-py==0.10.3
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
termcolor==1.1.0
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.17
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.1.9
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.3
